feat: add capability to join trees with shared roots
This came with a big refactor. First, variable naming was not the clearest,
so I addressed that before changing any logic.
Then, I added some comments to clarify the structure of the main method.
Following that, the Class approach seemed better for the Tree entity,
since I wanted to add a method to insert a child node at an arbitrary depth / point / node.